Implement synchronous driver flag
As per issue #197, NUT drivers work by default in asynchronous mode.
This means that all data are pushed by the driver on the communication
socket to upsd (Unix socket on Unix, Named pipe on Windows) without
waiting for these data to be actually consumed.  With some HW, such as
ePDUs, that can produce a lot of data, asynchronous mode may cause some
congestion, resulting in the socket to be full, and the driver to appear
as not connected.  By enabling the 'synchronous' flag, the driver will
wait for data to be consumed by upsd, prior to publishing more.  This
can be enabled either globally or per driver.