tripplitesu: Fix initialization when tripplite firmware is buggy
With some Tripplite SU1000RT2U (and possibly more) UPS's, a firmware bug causes
a malformed response to the very first command that is sent after the serial
port is opened following a warm or cold boot of the system. My theory is that
this related to either the RS232 data lines or handshaking lines being pulled
high once the server's UART is powered however I have not determined precisely
if this is related to the data line being pulled high or the handshaking lines
being asserted. However, I have been able to consistently reproduce the issue
where the driver fails to start on the first attempt after a cold/warm boot
across 3 different machines and 2 SU1000RT2U UPS's. To workaround this, the
initial enumeration is repeated a 2nd time after 300ms(to allow all garbage
data to arrive) if the first attempt fails, which allows the driver to
consistently startup successfully on the 1st attempt.

Closes networkupstools/nut#220